Shifted app grouping logic to utils file.
